requirements.txt
requests
feedparser
openai
Pillow
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
